[New Architecture!][Working infinite scrolling app][Slowness came from debugging][Async library handles generating new objects]
With the help of a Promise library, bolts-Android, the code now runs the same, but with async fetching in the background.

There is still a practical benefit of RoboSpice for making HTTP queries, but the code base is much simpler with the current setup.

Next step: Including an actual Neat Java library, and a CPPN activation library for creating picbreeder images! This is gun' be fun.