[Hooked up the "fake" genome outputs to generate the bitmaps on each display object]
This means that the process of generating new individuals is indeed happening on a background thread, then being loaded onto the Card UI when complete.

It would be a simple fix to swap in a NeatGenome and generate the pixels from a CPPN instead. That's the next step!

With dependency injection, it should be easy to swap in the new types, one piece at a time.