[Mwahahaha - a Picbreeder clone is alive!][Local evolution working][Single column display] [Like button working, impacts evolution][Next step: UI improvements + fetch from server]
There needs to be a little indication that the likes are influencing evolution. I'm looking at sticky headers as a way to indicate who the parents were for each individual. The cool thing would be that as you scroll, you can see parents for each individual (until you scroll to the next individual).

After the UI improvements, the next step would be fetching the artifacts from the server initially. After that, we'd need a front page view that can showcase the different artifacts. Finally, we'd hook up the publish button to the local server. Following that, we could test it on a live server.

win-save needs some updating as well. We'll need the concept of a user.